support PreprocessorConfig being dict in input
