+++ src/tabpfn/classifier.py
@@ -36,10 +36,10 @@ from tabpfn.base import (
 from tabpfn.constants import (
     PROBABILITY_EPSILON_ROUND_ZERO,
     SKLEARN_16_DECIMAL_PRECISION,
-    ModelInterfaceConfig,
     XType,
     YType,
 )
+from tabpfn.config import ModelInterfaceConfig
 from tabpfn.preprocessing import (
     ClassifierEnsembleConfig,
     EnsembleConfig,
